+++ server/handler/catalog/list.go
@@ -58,3 +58,19 @@ func ListArchs(c *gin.Context, repository db.AppRepository) {
 
 	c.JSON(http.StatusOK, gin.H{"archs": &archsList})
 }
+
+func ListPackages(c *gin.Context, repository db.AppRepository) {
+	ctx, ctxErr := context.WithTimeout(c.Request.Context(), 30*time.Second)
+	defer ctxErr()
+
+	var packageTypesList []*model.Package
+
+	//request on repository
+	if result, err := repository.ListPackages(ctx); err != nil {
+		logrus.Error(err)
+	} else {
+		packageTypesList = result
+	}
+
+	c.JSON(http.StatusOK, gin.H{"packages": &packageTypesList})
+}